DONE WITH CS410 BABY!!!
Made this new branch if I ever pick this back up. The sphere code is all really good.
It also contains the ray_cast_model function so you can do polygon intersection.

I don't have the color_me_poly code in here because it was so horrid, so for now
polys get set to a nice defauly cyan color